index: move server into app and add index.js
Most of the logic of server.js is now in app/server.js, but it doesn't
start anymore the server itself. It's now the index.js which does it.
The index.js is a more idiomatic starting point in the nodejs world.
And that way, we can do tests in a more programatic way.